+++ CHANGELOG.md
@@ -10,6 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - Created basic GLUT window for visual debugging
-- Implemented simple vertex shader
+- Implemented simple vertex and fragment shader
 
 [unreleased]: https://github.com/jdelacroix/glamm/compare/v0.1.0...HEAD